# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "4ks recipe API.\n\nAuthentication uses `Authorization: Bearer \u003ctoken\u003e`.\nMost authenticated routes expect an Auth0 JWT.\nApproved recipe routes documented as AI Kitchen Pass compatible also accept a Kitchen Pass bearer token.",
        "title": "4ks API",
        "contact": {},
        "version": "2.0"
    },
    "paths": {
...
        "/api/recipes/revisions/{revisionID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a Revision By ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get a Recipe Revision",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Revision ID",
                        "name": "revisionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RecipeRevision"
                        }
                    }
                }
            }
        },
...
        "/api/recipes/{recipeID}/fork": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fork Recipe. This route accepts either an Auth0 JWT or an AI Kitchen Pass bearer token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Fork Recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Recipe"
                        }
                    }
                }
            }
        },
...
        "/api/recipes/{recipeID}/revisions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all revisions for a Recipe. This route is part of the AI Kitchen Pass workflow surface.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Get all revisions for a Recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RecipeRevision"
                            }
                        }
                    }
                }
            }
        },
        "/api/recipes/{recipeID}/star": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Star Recipe",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Recipes"
                ],
                "summary": "Star Recipe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recipe ID",
                        "name": "recipeID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Ingredient": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Instruction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.Recipe": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.UserSummary"
                },
                "branch": {
                    "type": "string"
                },
                "contributors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.UserSummary"
                    }
                },
                "createdDate": {
                    "type": "string"
                },
                "currentRevision": {
                    "$ref": "#/definitions/models.RecipeRevision"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.RecipeMetadata"
                },
                "root": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                }
            }
        },
...
        "models.RecipeMediaVariant": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "maxWidth": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.RecipeMetadata": {
            "type": "object",
            "properties": {
                "forks": {
                    "type": "integer"
                },
                "stars": {
                    "type": "integer"
                }
            }
        },
        "models.RecipeRevision": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/models.UserSummary"
                },
                "banner": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.RecipeMediaVariant"
                    }
                },
                "createdDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Ingredient"
                    }
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instruction"
                    }
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "recipeId": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                }
            }
        },
...
        "models.UserSummary": {
            "type": "object",
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}